Fixing bug in the code. Making the weights parameter to be set on the command line, to avoid generating too much cfg files.

+++ research/novelty/on-policy/control/RLLearner.hpp
@@ -15,7 +15,7 @@ class RLLearner : public Agent{
 	protected:
 		ActionVect actions;
 
-		vector<int> option;
+		vector<double> option;
 		string pathToRewardDescription;
 
 		double gamma, epsilon;
